move club entity into its own package; update references
