+++ dashboard-server/src/main/java/com/footballstatsdashboard/api/model/club/Club.java
@@ -1,4 +1,4 @@
-package com.footballstatsdashboard.api.model;
+package com.footballstatsdashboard.api.model.club;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonInclude;
